Moved the export of a component from parent to only child that actually uses the component
